Fix checking of parent presence in namespace method
The first_auto_devops_config method checks for the presence
of a `parent` using the `has_parent?` function. `has_parent?`
checks first for a `parent_id`, this causes an error when
parent_id is still present but the actual parent object was
deleted.

This is just a quick fix to unblock an ongoing error in
deployed instances.

Changelog: fixed