Add the UserType in create/remove addon response
We need to be able to update the UI with the user state after an add on
has been assigned/de-assigned. Add the user to the mutation responses so
that the frontend can query for this when needed.

Changelog: added
MR: https://gitlab.com/gitlab-org/gitlab/-/merge_requests/129887
Issue: https://gitlab.com/gitlab-org/gitlab/-/issues/421974
EE: true