Remove cache_pages_domain_api feature flag
Although the cache improves the API performance, it's quite flaky and
sensitive to too many events, which makes it too hard to maintain.

We enabled the feature flag a few times, but each time it cause a
different incident (examples: [1], [2]).

Since other works were done to [improve the performance], we decided to
not continue with the feature flag.

Related to: https://gitlab.com/gitlab-org/gitlab/-/issues/425143

[improve the performance]: https://gitlab.com/gitlab-org/gitlab/-/issues/378508
[1]: https://gitlab.com/gitlab-org/gitlab-pages/-/issues/827#note_1262875898
[2]: https://gitlab.com/gitlab-org/gitlab/-/issues/373220#note_1205506981

Changelog: removed