Deduplicate advisory scans for the same advisory
It's expected for advisory scans to be a long running job, and
indempotent. If multiple scans for the same advisory run, the same set
of vulnerabilities will have been created making the computational
resources used on a no-op. This change prevents this and lessens the
load of advisory scans on other Sidekiq workers.

Co-authored-by: Brian Williams <bwilliams@gitlab.com>