Add PagesDeployment#deleted_at field
Currently when a new `PagesDeployment` is created, we add its id to
`ProjectPagesMetadatum`. This step is required as the new deployment
might take a while to be uploaded to the Object Storage, therefore,
instead of only replacing the current deployment for the project, we
create a new one to avoid race conditions.

With the work to introduce [Gitlab Pages Multiple Versions], a project
now have to support multiple `PagesDeployment` at the same time. For
this reason, instead of using
`ProjectPagesMetadatum#pages_deployment_id` to identify _valid_
deployment, we're introducing a `PagesDeployment#deleted_at` field.

The `PagesDeployment#deleted_at` is used to signalize when a new
deployment for a project was created, marking old ones with `deleted_at`
date. This will be used to clean _old_ deployments and filter for
_active_ deployments.

In the near future it'll differentiate when a new deployment for a given
project and [path_prefix] was created.

Related to:
- https://gitlab.com/gitlab-org/gitlab/-/issues/328715+
- https://gitlab.com/gitlab-org/gitlab/-/issues/416494+

[Gitlab Pages Multiple Versions]: https://gitlab.com/groups/gitlab-org/-/epics/10914
[path_prefix]: https://gitlab.com/gitlab-org/gitlab/-/merge_requests/129534

Changelog: changed