Enforce no transactions on specs with delete tables strategy
We know using testprof method like let_it_be and before_all
don't mix well with specs that are supposed to run outside
of a transaction (since they create one).
This change raises an error whenver a transaction is open
on these type of specs so the developer can fix it.