Don't signal MR rebase if on current fast-forward train
Currently, when the first car of a fast-forward merge train is merged,
MRs for subsequent cars on the train are shown as "should be rebased" in
the UI.

But this is not necessary, because fast-forward merge trains
automatically rebase internally on their merge train ref, and perform a
fast-forward merge directly from the train ref. And if the train ref
turns out to be out of date, then the merge itself will fail later the
fast-forward condition would be broken.

See https://gitlab.com/gitlab-org/gitlab/-/issues/420000

Changelog: fixed
EE: true