Move repository storage to BasicProjectDetails
What
---
Move the `repository_storage_detail` from the project entities to the
basic project details.

Why
---
[`forked_from_project`](https://gitlab.com/gitlab-org/gitlab/-/blob/b68c62699fabe878e081ed3f64c6fea412ee4f2e/lib/api/entities/project.rb#L94)
returns `BasicPrjectDetails` which doesn't include the repository
storage. Having the repository storage available will allow to easily
know if the project and the `forked_from_project` are in the same
storage or not so that we can move them to the same storage.

We can still get this information if we send a `GET /project/:id`
request but this will be en extra API call for the user every time they
want to see if a project and it's root project (where its forked from)
are in the same gitaly shard.

Testing
---

1. Start a fresh GDK: `gdk reset-data`.
1. Create a new PAT with `api` scope for the `root` user.
1. Send requests to a project that is a fork:
    Before:

    ```shell
    $ curl -s --request GET --header "PRIVATE-TOKEN: $ADMIN_PAT" "http://gdk.test:3000/api/v4/projects/bethanie%2FFlight" | jq -r '.forked_from_project.repository_storage'
    null
    ```

    After:

    ```shell
    $ curl -s --request GET --header "PRIVATE-TOKEN: $ADMIN_PAT" "http://gdk.test:3000/api/v4/projects/bethanie%2FFlight" | jq -r '.forked_from_project.repository_storage'
    default
    ```

1. Impersonate a normal user from <http://gdk.test:3000/admin/users/>
1. Create a new PAT with `api` scope.
1. Send a request to a project that is a fork:

    ```shell
    # Check that repository_storage is not returned
    $ curl -s --request GET --header "PRIVATE-TOKEN: $USER_PAT" "http://gdk.test:3000/api/v4/projects/bethanie%2FFlight" | jq -r '.forked_from_project.repository_storage'
    null
    # Validate that `forked_from_project` is still returned.
    $ curl -s --request GET --header "PRIVATE-TOKEN: $USER_PAT" "http://gdk.test:3000/api/v4/projects/bethanie%2FFlight" | jq -r '.forked_from_project.id'
    17
    ```

Changelog: added
Reference: https://gitlab.com/gitlab-com/gl-infra/reliability/-/issues/24322
Signed-off-by: Steve Xuereb <sxuereb@gitlab.com>