Improve VertexAI embeddings creation performance
This MR changes the flow of the embeddings creation as follows:
- md5sum on the file content is computed and stored as metadata
on DB embedding record.
- CreateEmptyEmbeddingsRecordsWorker now checks if the md5sum for
contents of a file changed before re-creating the embeddings for
that file
- If file contents are different CreateEmptyEmbeddingsRecordsWorker
will enqueue a CreateDbEmbeddingsPerDocFileWorker per filea and
new version, which will allow deduplication of jobs to take place for
CreateDbEmbeddingsPerDocFileWorker in case CreateEmptyEmbeddingsRecordsWorker
failed and is being rerun.
- CreateDbEmbeddingsPerDocFileWorker on re-run checks for any existing
embeddings for the given file at given version and cleans-up those first
before creating new embedding records, which prevents duplications.