Document how to revoke GitLab agent token
What
---
Update the token overview page with GitLab agent token information, and
how to revoke.

Why
---
When you [register GitLab
agent](https://docs.gitlab.com/ee/user/clusters/agent/install/#register-the-agent-with-gitlab)
an access token is created which has read-only access to the project
repository. We don't provide a lot of information on how to find the token
information and how to revoke it.

Signed-off-by: Steve Xuereb <sxuereb@gitlab.com>