Respond with 401 instead of 404 on npm unauthenticated requests
At present, if an unauthenticated request is made to a private package, the server responds with a 404. This could be confusing to users, since they may assume that the package does not exist, rather than focus on the fact that the request is not authenticated. In this MR, this behavior is changed to respond with a 401 instead.

Changelog: changed