Drop unused test run script for attr_encrypted
We vendored this gem a while ago, and it contained
a lot of code to deal with older versions of Ruby
and Rails that are not relevant to use.

This test run script is not used either, we just
call straight into rspec in CI/CD via bundler.

Let's drop this since it's leading to issue reports.