Broadcast chat messages to all clients
This change introduces the `aiAction` parameter as part of the
`aiCompletion` subscription. It currently only works for chat and is
used to receive all chat responses from a user over the subscription.
This enables us to sync all open chats from a client.

To not break existing chat clients, we now broadcast events twice:
1. For old clients via `userId`, `resourceId`
2. For new clients via `userId`, `aiAction`

We can remove broadcasting twice once all clients are up to date.

We also use the "streamed" subscription to broadcast the currently
used tool. Otherwise, new clients that only listen on `userId` and
`aiAction` would not receive the current tool (as this is a response
that should be broadcasted to all clients).
Tools now get broadcasted using `userId` and `clientSubscriptionId`.

Changelog: added
EE: true