Rescue NoRepository exceptions when fetching patch_id
This error is returned when either old_revision or new_revision do
not exist in the repository. Similar to Gitlab::Git::CommandError, we
want to return nil, as it is generally a non-fatal exception, so we
should exit gracefully.

Additionally, we're adding logging of both exceptions, so that we can
track the incident rate, and will help with debugging customer issues.