Create RefExtractor class
RefExtractor is a refactor of ExtractsRef module.

ExtractsRef was mostly used as a Rails concern in controllers
to extract git related strings from request parameters.
The module utilized instance variables that made the code
hard to follow and difficult to examine for side effect.

As of this commit, ExtractsRef cannot be removed yet because
another related module ExtractsPath depends on ExtractsRef.