Ensure the train ref is mergeable
In order for it to be safe to fast-forward merge directly from the merge
train ref, the ref needs to look exactly like the target branch would on
merge. This was not always the case: Old merge trains did not respect
most merge preferences, such as the commit message, squashing, and the
project's merge method. Thus, it was possible for train refs to include
commit messages that should not be merged.

To solve this, we rely on the presence of a recently added merge parameter
and ensure it is only present for mergeable train refs.

https://gitlab.com/gitlab-org/gitlab/-/issues/421588